# Order rows during incremental index persist when rollup is disabled.