--- a/processing/src/main/java/io/druid/segment/incremental/IncrementalIndex.java
+++ b/processing/src/main/java/io/druid/segment/incremental/IncrementalIndex.java
@@ -80,6 +80,7 @@ import javax.annotation.concurrent.GuardedBy;
 import java.io.Closeable;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Comparator;
 import java.util.Deque;
 import java.util.Iterator;
@@ -94,6 +95,7 @@ import java.util.concurrent.ConcurrentSkipListMap;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
+import java.util.stream.Stream;
 
 /**
  */
@@ -368,6 +370,16 @@ public abstract class IncrementalIndex<AggregatorType> extends AbstractIndex imp
     {
       this.incrementalIndexSchema = new IncrementalIndexSchema.Builder()
           .withMetrics(metrics)
+          .build();
+      return this;
+    }
+
+    @VisibleForTesting
+    public Builder setSimpleTestingIndexSchema(boolean rollup, final AggregatorFactory... metrics)
+    {
+      this.incrementalIndexSchema = new IncrementalIndexSchema.Builder()
+          .withMetrics(metrics)
+          .withRollup(rollup)
           .build();
       return this;
     }
@@ -1203,6 +1215,12 @@ public abstract class IncrementalIndex<AggregatorType> extends AbstractIndex imp
     Iterable<IncrementalIndexRow> keySet();
 
     /**
+     * Get all {@link IncrementalIndexRow} to persist, ordered with {@link Comparator<IncrementalIndexRow>}
+     * @return
+     */
+    Iterable<IncrementalIndexRow> persistIterable();
+
+    /**
      * @return the previous rowIndex associated with the specified key, or
      * {@link IncrementalIndexRow#EMPTY_ROW_INDEX} if there was no mapping for the key.
      */
@@ -1290,6 +1308,13 @@ public abstract class IncrementalIndex<AggregatorType> extends AbstractIndex imp
     }
 
     @Override
+    public Iterable<IncrementalIndexRow> persistIterable()
+    {
+      // with rollup, facts are already pre-sorted so just return keyset
+      return keySet();
+    }
+
+    @Override
     public int putIfAbsent(IncrementalIndexRow key, int rowIndex)
     {
       // setRowIndex() must be called before facts.putIfAbsent() for visibility of rowIndex from concurrent readers.
@@ -1310,7 +1335,9 @@ public abstract class IncrementalIndex<AggregatorType> extends AbstractIndex imp
     private final boolean sortFacts;
     private final ConcurrentMap<Long, Deque<IncrementalIndexRow>> facts;
 
-    public PlainFactsHolder(boolean sortFacts)
+    private final Comparator<IncrementalIndexRow> incrementalIndexRowComparator;
+
+    public PlainFactsHolder(boolean sortFacts, Comparator<IncrementalIndexRow> incrementalIndexRowComparator)
     {
       this.sortFacts = sortFacts;
       if (sortFacts) {
@@ -1318,6 +1345,7 @@ public abstract class IncrementalIndex<AggregatorType> extends AbstractIndex imp
       } else {
         this.facts = new ConcurrentHashMap<>();
       }
+      this.incrementalIndexRowComparator = incrementalIndexRowComparator;
     }
 
     @Override
@@ -1351,10 +1379,10 @@ public abstract class IncrementalIndex<AggregatorType> extends AbstractIndex imp
     public Iterator<IncrementalIndexRow> iterator(boolean descending)
     {
       if (descending && sortFacts) {
-        return concat(((ConcurrentNavigableMap<Long, Deque<IncrementalIndexRow>>) facts)
+        return timeOrderedConcat(((ConcurrentNavigableMap<Long, Deque<IncrementalIndexRow>>) facts)
                 .descendingMap().values(), true).iterator();
       }
-      return concat(facts.values(), false).iterator();
+      return timeOrderedConcat(facts.values(), false).iterator();
     }
 
     @Override
@@ -1363,10 +1391,10 @@ public abstract class IncrementalIndex<AggregatorType> extends AbstractIndex imp
       ConcurrentNavigableMap<Long, Deque<IncrementalIndexRow>> subMap =
           ((ConcurrentNavigableMap<Long, Deque<IncrementalIndexRow>>) facts).subMap(timeStart, timeEnd);
       final Map<Long, Deque<IncrementalIndexRow>> rangeMap = descending ? subMap.descendingMap() : subMap;
-      return concat(rangeMap.values(), descending);
+      return timeOrderedConcat(rangeMap.values(), descending);
     }
 
-    private Iterable<IncrementalIndexRow> concat(
+    private Iterable<IncrementalIndexRow> timeOrderedConcat(
         final Iterable<Deque<IncrementalIndexRow>> iterable,
         final boolean descending
     )
@@ -1379,10 +1407,31 @@ public abstract class IncrementalIndex<AggregatorType> extends AbstractIndex imp
       );
     }
 
+    private Stream<IncrementalIndexRow> timeAndDimsOrderedConcat(
+        final Collection<Deque<IncrementalIndexRow>> rowGroups,
+        final boolean descending
+    )
+    {
+      if (descending) {
+        return rowGroups.stream()
+                        .flatMap(Collection::stream)
+                        .sorted(incrementalIndexRowComparator.reversed());
+      }
+      return rowGroups.stream()
+                      .flatMap(Collection::stream)
+                      .sorted(incrementalIndexRowComparator);
+    }
+
     @Override
     public Iterable<IncrementalIndexRow> keySet()
     {
-      return concat(facts.values(), false);
+      return timeOrderedConcat(facts.values(), false);
+    }
+
+    @Override
+    public Iterable<IncrementalIndexRow> persistIterable()
+    {
+      return () -> timeAndDimsOrderedConcat(facts.values(), false).iterator();
     }
 
     @Override