--- a/processing/src/test/java/io/druid/segment/incremental/IncrementalIndexAdapterTest.java
+++ b/processing/src/test/java/io/druid/segment/incremental/IncrementalIndexAdapterTest.java
@@ -19,6 +19,7 @@
 
 package io.druid.segment.incremental;
 
+import io.druid.java.util.common.StringUtils;
 import io.druid.segment.IndexSpec;
 import io.druid.segment.IndexableAdapter;
 import io.druid.segment.RowIterator;
@@ -32,6 +33,7 @@ import org.junit.Test;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.function.Function;
 
 public class IncrementalIndexAdapterTest
 {
@@ -82,5 +84,82 @@ public class IncrementalIndexAdapterTest
     Assert.assertEquals(2, rowNums.size());
     Assert.assertEquals(0, (long) rowNums.get(0));
     Assert.assertEquals(1, (long) rowNums.get(1));
+  }
+
+  @Test
+  public void testGetRowsIterableNoRollup() throws Exception
+  {
+    final long timestamp = System.currentTimeMillis();
+    IncrementalIndex toPersist1 = IncrementalIndexTest.createNoRollupIndex(null);
+    IncrementalIndexTest.populateIndex(timestamp, toPersist1);
+    IncrementalIndexTest.populateIndex(timestamp, toPersist1);
+    IncrementalIndexTest.populateIndex(timestamp, toPersist1);
+
+
+    ArrayList<Integer> dim1Vals = new ArrayList<>();
+    for (IncrementalIndexRow row : toPersist1.getFacts().keySet()) {
+      dim1Vals.add(((int[]) row.getDims()[0])[0]);
+    }
+    ArrayList<Integer> dim2Vals = new ArrayList<>();
+    for (IncrementalIndexRow row : toPersist1.getFacts().keySet()) {
+      dim2Vals.add(((int[]) row.getDims()[1])[0]);
+    }
+
+    final IndexableAdapter incrementalAdapter = new IncrementalIndexAdapter(
+        toPersist1.getInterval(),
+        toPersist1,
+        INDEX_SPEC.getBitmapSerdeFactory()
+                  .getBitmapFactory()
+    );
+
+    RowIterator rows = incrementalAdapter.getRows();
+    List<String> rowStrings = new ArrayList<>();
+    while (rows.moveToNext()) {
+      rowStrings.add(rows.getPointer().toString());
+    }
+
+    Function<Integer, String> getExpected = (rowNumber) -> {
+      if (rowNumber < 3) {
+        return StringUtils.format(
+            "RowPointer{indexNum=0, rowNumber=%s, timestamp=%s, dimensions={dim1=1, dim2=2}, metrics={count=1}}",
+            rowNumber,
+            timestamp
+        );
+      } else {
+        return StringUtils.format(
+            "RowPointer{indexNum=0, rowNumber=%s, timestamp=%s, dimensions={dim1=3, dim2=4}, metrics={count=1}}",
+            rowNumber,
+            timestamp
+        );
+      }
+    };
+
+
+    // without sorting, output would be
+    //    RowPointer{indexNum=0, rowNumber=0, timestamp=1533347274588, dimensions={dim1=1, dim2=2}, metrics={count=1}}
+    //    RowPointer{indexNum=0, rowNumber=1, timestamp=1533347274588, dimensions={dim1=3, dim2=4}, metrics={count=1}}
+    //    RowPointer{indexNum=0, rowNumber=2, timestamp=1533347274588, dimensions={dim1=1, dim2=2}, metrics={count=1}}
+    //    RowPointer{indexNum=0, rowNumber=3, timestamp=1533347274588, dimensions={dim1=3, dim2=4}, metrics={count=1}}
+    //    RowPointer{indexNum=0, rowNumber=4, timestamp=1533347274588, dimensions={dim1=1, dim2=2}, metrics={count=1}}
+    //    RowPointer{indexNum=0, rowNumber=5, timestamp=1533347274588, dimensions={dim1=3, dim2=4}, metrics={count=1}}
+    // but with sorting, output should be
+    //    RowPointer{indexNum=0, rowNumber=0, timestamp=1533347361396, dimensions={dim1=1, dim2=2}, metrics={count=1}}
+    //    RowPointer{indexNum=0, rowNumber=1, timestamp=1533347361396, dimensions={dim1=1, dim2=2}, metrics={count=1}}
+    //    RowPointer{indexNum=0, rowNumber=2, timestamp=1533347361396, dimensions={dim1=1, dim2=2}, metrics={count=1}}
+    //    RowPointer{indexNum=0, rowNumber=3, timestamp=1533347361396, dimensions={dim1=3, dim2=4}, metrics={count=1}}
+    //    RowPointer{indexNum=0, rowNumber=4, timestamp=1533347361396, dimensions={dim1=3, dim2=4}, metrics={count=1}}
+    //    RowPointer{indexNum=0, rowNumber=5, timestamp=1533347361396, dimensions={dim1=3, dim2=4}, metrics={count=1}}
+
+    Assert.assertEquals(6, rowStrings.size());
+    for (int i = 0; i < 6; i++) {
+      if (i % 2 == 0) {
+        Assert.assertEquals(0, (long) dim1Vals.get(i));
+        Assert.assertEquals(0, (long) dim2Vals.get(i));
+      } else {
+        Assert.assertEquals(1, (long) dim1Vals.get(i));
+        Assert.assertEquals(1, (long) dim2Vals.get(i));
+      }
+      Assert.assertEquals(getExpected.apply(i), rowStrings.get(i));
+    }
   }
 }