@gianm 2018-08-06
Tests look good now, so I'll merge it. Thanks @clintropolis!